+++ sdk/core/src/data/transaction/api_transaction_result.rs
