+++ sdk/core/src/data/transaction/get_transaction_response.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use super::transaction_on_network::ApiTransactionResult;
+use super::api_transaction_result::ApiTransactionResult;
 
 /// Corresponds to [`GetTransactionResponseData`](https://github.com/multiversx/mx-chain-proxy-go/blob/main/data/transaction.go) in mx-chain-proxy-go.
 #[derive(Debug, Clone, Serialize, Deserialize)]
