+++ sdk/http/src/gateway_http_proxy/http_tx.rs
@@ -4,7 +4,7 @@ use multiversx_sdk::{
     data::{
         network_config::NetworkConfig,
         transaction::{
-            ArgCreateTransaction, Transaction, ApiTransactionResult, TxCostResponseData,
+            ApiTransactionResult, ArgCreateTransaction, Transaction, TxCostResponseData,
         },
         vm::{VMQueryInput, VmValuesResponseData},
     },
